requirements.txt
# Core dependencies
boto3>=1.28.0
pandas>=2.0.0
pyyaml>=6.0
click>=8.0.0
rich>=13.0.0
python-dotenv>=1.0.0

langchain>=0.1.0,<1.0.0
langchain-aws>=0.1.0
jinja2>=3.0.0
